backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.2.0
black==23.11.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.9
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-timezone-field==6.1.0
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
drf-spectacular==0.27.0
flower==2.0.1
humanize==4.9.0
inflection==0.5.1
install==1.3.5
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kombu==5.3.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.0
rpds-py==0.15.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
